split tests according to subspecs functionality
